% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{lens}
\alias{lens}
\title{Slice an image and reorder it twice}
\usage{
lens(img, dpi = 30)
}
\arguments{
\item{img}{magick image object returned by image_read() or image_graph()}

\item{dpi}{number of slice}
}
\description{
Slice an image and reorder it twice
}
\examples{
library(magick)
image_read("http://fr.web.img6.acsta.net/pictures/15/07/20/18/14/582462.jpg") \%>\%
lens(dpi = 50)
}
