% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{slice_and_mix}
\alias{slice_and_mix}
\title{Slice an image and reorder it}
\usage{
slice_and_mix(img, dpi = 40)
}
\arguments{
\item{img}{magick image object returned by image_read() or image_graph()}

\item{dpi}{number of slice}
}
\description{
Slice an image and reorder it
}
\examples{
library(magick)
image_read("http://fr.web.img6.acsta.net/pictures/15/07/20/18/14/582462.jpg") \%>\%
slice_and_mix(dpi = 50)
}
